pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=2.13.0",
    "pillow>=12.0.0",
    "reportlab>=4.4.6",
    "stripe>=14.0.1",
    "werkzeug>=3.1.4",
]
